+++ pypdf/filters.py
@@ -37,17 +37,14 @@ __author_email__ = "biziqe@mathieu.fenniak.net"
 import math
 import struct
 import zlib
-from io import BytesIO
-from typing import Any, Dict, List, Optional, Tuple, Union, cast
+from typing import Any, List, Union, cast
 
 from ._utils import (
     b_,
     deprecate_with_replacement,
-    logger_warning,
     ord_,
 )
 from .constants import CcittFaxDecodeParameters as CCITT
-from .constants import ColorSpaces
 from .constants import FilterTypeAbbreviations as FTA
 from .constants import FilterTypes as FT
 from .constants import ImageAttributes as IA
@@ -715,174 +712,3 @@ def decodeStreamData(stream: Any) -> Union[str, bytes]:  # deprecated
     """Deprecated. Use decode_stream_data."""
     deprecate_with_replacement("decodeStreamData", "decode_stream_data", "4.0.0")
     return decode_stream_data(stream)
-
-
-def _xobj_to_image(x_object_obj: Dict[str, Any]) -> Tuple[Optional[str], bytes, Any]:
-    """
-    Users need to have the pillow package installed.
-
-    It's unclear if pypdf will keep this function here, hence it's private.
-    It might get removed at any point.
-
-    Args:
-      x_object_obj:
-
-    Returns:
-        Tuple[file extension, bytes, PIL.Image.Image]
-    """
-    from ._xobj_image_helpers import (
-        Image,
-        _get_imagemode,
-        _handle_flate,
-        _handle_jpx,
-        mode_str_type,
-    )
-
-    # for error reporting
-    if (
-        hasattr(x_object_obj, "indirect_reference") and x_object_obj is None
-    ):  # pragma: no cover
-        obj_as_text = x_object_obj.indirect_reference.__repr__()
-    else:
-        obj_as_text = x_object_obj.__repr__()
-
-    size = (x_object_obj[IA.WIDTH], x_object_obj[IA.HEIGHT])
-    data = x_object_obj.get_data()  # type: ignore
-    if isinstance(data, str):  # pragma: no cover
-        data = data.encode()
-    colors = x_object_obj.get("/Colors", 1)
-    color_space: Any = x_object_obj.get("/ColorSpace", NullObject()).get_object()
-    if isinstance(color_space, list) and len(color_space) == 1:
-        color_space = color_space[0].get_object()
-    if (
-        IA.COLOR_SPACE in x_object_obj
-        and x_object_obj[IA.COLOR_SPACE] == ColorSpaces.DEVICE_RGB
-    ):
-        # https://pillow.readthedocs.io/en/stable/handbook/concepts.html#modes
-        mode: mode_str_type = "RGB"
-    if x_object_obj.get("/BitsPerComponent", 8) < 8:
-        mode, invert_color = _get_imagemode(
-            f"{x_object_obj.get('/BitsPerComponent', 8)}bit", 0, ""
-        )
-    else:
-        mode, invert_color = _get_imagemode(
-            color_space,
-            2
-            if (
-                colors == 1
-                and (
-                    not isinstance(color_space, NullObject)
-                    and "Gray" not in color_space
-                )
-            )
-            else colors,
-            "",
-        )
-    extension = None
-    alpha = None
-    filters = x_object_obj.get(SA.FILTER, [None])
-    lfilters = filters[-1] if isinstance(filters, list) else filters
-    if lfilters == FT.FLATE_DECODE:
-        img, image_format, extension, invert_color = _handle_flate(
-            size,
-            data,
-            mode,
-            color_space,
-            colors,
-            obj_as_text,
-        )
-    elif lfilters in (FT.LZW_DECODE, FT.ASCII_85_DECODE, FT.CCITT_FAX_DECODE):
-        # I'm not sure if the following logic is correct.
-        # There might not be any relationship between the filters and the
-        # extension
-        if x_object_obj[SA.FILTER] in [[FT.LZW_DECODE], [FT.CCITT_FAX_DECODE]]:
-            extension = ".tiff"  # mime_type = "image/tiff"
-            image_format = "TIFF"
-        else:
-            extension = ".png"  # mime_type = "image/png"
-            image_format = "PNG"
-        img = Image.open(BytesIO(data), formats=("TIFF", "PNG"))
-    elif lfilters == FT.DCT_DECODE:
-        img, image_format, extension = Image.open(BytesIO(data)), "JPEG", ".jpg"
-        # invert_color kept unchanged
-    elif lfilters == FT.JPX_DECODE:
-        img, image_format, extension, invert_color = _handle_jpx(
-            size, data, mode, color_space, colors
-        )
-    elif lfilters == FT.CCITT_FAX_DECODE:
-        img, image_format, extension, invert_color = (
-            Image.open(BytesIO(data), formats=("TIFF",)),
-            "TIFF",
-            ".tiff",
-            False,
-        )
-    else:
-        if mode == "":
-            raise PdfReadError(f"ColorSpace field not found in {x_object_obj}")
-        img, image_format, extension, invert_color = (
-            Image.frombytes(mode, size, data),
-            "PNG",
-            ".png",
-            False,
-        )
-
-    # CMYK image and other colorspaces without decode
-    # requires reverting scale (cf p243,2§ last sentence)
-    decode = x_object_obj.get(
-        IA.DECODE,
-        ([1.0, 0.0] * len(img.getbands()))
-        if (
-            (img.mode == "CMYK" or (invert_color and img.mode == "L"))
-            and lfilters in (FT.DCT_DECODE, FT.JPX_DECODE)
-        )
-        else None,
-    )
-    if (
-        isinstance(color_space, ArrayObject)
-        and color_space[0].get_object() == "/Indexed"
-    ):
-        decode = None  # decode is meanless of Indexed
-    if decode is not None and not all(decode[i] == i % 2 for i in range(len(decode))):
-        lut: List[int] = []
-        for i in range(0, len(decode), 2):
-            dmin = decode[i]
-            dmax = decode[i + 1]
-            lut.extend(
-                round(255.0 * (j / 255.0 * (dmax - dmin) + dmin)) for j in range(256)
-            )
-        img = img.point(lut)
-
-    if IA.S_MASK in x_object_obj:  # add alpha channel
-        alpha = _xobj_to_image(x_object_obj[IA.S_MASK])[2]
-        if img.size != alpha.size:
-            logger_warning(f"image and mask size not matching: {obj_as_text}", __name__)
-        else:
-            # TODO : implement mask
-            if alpha.mode != "L":
-                alpha = alpha.convert("L")
-            if img.mode == "P":
-                img = img.convert("RGB")
-            elif img.mode == "1":
-                img = img.convert("L")
-            img.putalpha(alpha)
-        if "JPEG" in image_format:
-            extension = ".jp2"
-            image_format = "JPEG2000"
-        else:
-            extension = ".png"
-            image_format = "PNG"
-
-    img_byte_arr = BytesIO()
-    try:
-        img.save(img_byte_arr, format=image_format)
-    except OSError:  # pragma: no cover
-        # odd error
-        img_byte_arr = BytesIO()
-        img.save(img_byte_arr, format=image_format)
-    data = img_byte_arr.getvalue()
-
-    try:  # temporary try/except until other fixes of images
-        img = Image.open(BytesIO(data))
-    except Exception:
-        img = None  # type: ignore
-    return extension, data, img
